requirements.txt
playwright~=1.58.0
requests~=2.32.5
python-dotenv

mysql-connector-python~=9.6.0